# openapi_gpt.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hepatology Research OS (GPT Surface)",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "https://clinicaltrialapi-dqml.onrender.com"
    }
  ],
  "paths": {
...
    "/protocol/endpoints": {
      "post": {
        "tags": [
          "Protocol Development"
        ],
        "summary": "Generate Protocol Endpoints",
        "operationId": "generate_protocol_endpoints_protocol_endpoints_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProtocolEndpointsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/protocol/estimand": {
      "post": {
        "tags": [
          "Protocol Development"
        ],
        "summary": "Generate Estimand",
        "operationId": "generate_estimand_protocol_estimand_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProtocolEstimandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sap/statistical-analysis-plan": {
      "post": {
        "tags": [
          "Statistical Analysis Plan"
        ],
        "summary": "Generate Sap",
        "operationId": "generate_sap_sap_statistical_analysis_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SAPRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProtocolEndpointsRequest": {
        "properties": {
          "disease": {
            "type": "string",
            "title": "Disease"
          },
          "intervention": {
            "type": "string",
            "title": "Intervention"
          },
          "primary_objective": {
            "type": "string",
            "title": "Primary Objective"
          }
        },
        "type": "object",
        "required": [
          "disease",
          "intervention",
          "primary_objective"
        ],
        "title": "ProtocolEndpointsRequest"
      },
      "ProtocolEstimandRequest": {
        "properties": {
          "treatment": {
            "type": "string",
            "title": "Treatment"
          },
          "population": {
            "type": "string",
            "title": "Population"
          },
          "endpoint": {
            "type": "string",
            "title": "Endpoint"
          }
        },
        "type": "object",
        "required": [
          "treatment",
          "population",
          "endpoint"
        ],
        "title": "ProtocolEstimandRequest"
      },
...
      "SAPRequest": {
        "properties": {
          "endpoint_type": {
            "type": "string",
            "title": "Endpoint Type"
          },
          "endpoint_name": {
            "type": "string",
            "title": "Endpoint Name"
          }
        },
        "type": "object",
        "required": [
          "endpoint_type",
          "endpoint_name"
        ],
        "title": "SAPRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}